`StegrWork` as a separate package
